requirements.txt
sqlparse==0.2.4
Django==2.2.11
djongo==1.3.1
pillow==7.0.0
djangorestframework==3.11.0
django-rest-auth==0.9.5
django-allauth==0.41.0
flake8